requirements.txt
# Core Data Engineering
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

# Spotify API
spotipy>=2.22.0
requests>=2.28.0

# Data Analytics
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.11.0
streamlit>=1.13.0

# Data Science
scikit-learn>=1.1.0
tensorflow>=2.10.0

# Workflow & Scheduling
apache-airflow>=3.1.0
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-celery>=3.0.0
apache-airflow-providers-redis>=3.0.0

# Development & Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
jupyter>=1.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
loguru>=0.6.0
fastapi>=0.68.0
uvicorn>=0.15.0

# Redis (optional)
redis>=4.0.0

# Additional Docker-friendly packages
gunicorn>=20.1.0
psutil>=5.8.0

# We'll add more packages in later days
# streamlit (Day 8)
# scikit-learn (Day 15)
# plotly (Day 11)

